refactor Loginfilters, merge conflicts
