+++ backend/src/main/java/com/beyond/specguard/admin/controller/AdminAuthController.java
@@ -4,8 +4,8 @@ import com.beyond.specguard.admin.model.dto.request.InternalAdminRequestDto;
 import com.beyond.specguard.admin.model.dto.response.InternalAdminResponseDto;
 import com.beyond.specguard.admin.model.service.InternalAdminService;
 import com.beyond.specguard.auth.model.dto.response.ReissueResponseDto;
-import com.beyond.specguard.auth.model.service.LogoutService;
-import com.beyond.specguard.auth.model.service.ReissueService;
+import com.beyond.specguard.auth.model.service.common.LogoutService;
+import com.beyond.specguard.auth.model.service.common.ReissueService;
 import com.beyond.specguard.common.util.CookieUtil;
 import com.beyond.specguard.common.util.TokenResponseWriter;
 import io.swagger.v3.oas.annotations.Operation;
