+++ backend/src/main/java/com/beyond/specguard/auth/controller/AuthController.java
@@ -5,11 +5,12 @@ import com.beyond.specguard.auth.model.dto.request.SignupRequestDto;
 import com.beyond.specguard.auth.model.dto.response.InviteCheckResponseDto;
 import com.beyond.specguard.auth.model.dto.response.ReissueResponseDto;
 import com.beyond.specguard.auth.model.dto.response.SignupResponseDto;
-import com.beyond.specguard.auth.model.service.InviteSignupService;
-import com.beyond.specguard.auth.model.service.LogoutService;
-import com.beyond.specguard.auth.model.service.ReissueService;
-import com.beyond.specguard.auth.model.service.SignupService;
-import com.beyond.specguard.common.jwt.JwtUtil;
+import com.beyond.specguard.auth.model.dto.response.TokenResponseDto;
+import com.beyond.specguard.auth.model.service.common.LogoutService;
+import com.beyond.specguard.auth.model.service.common.ReissueService;
+import com.beyond.specguard.auth.model.service.local.InviteSignupService;
+import com.beyond.specguard.auth.model.service.local.SignupService;
+import com.beyond.specguard.auth.model.service.oauth2.TokenService;
 import com.beyond.specguard.common.util.CookieUtil;
 import com.beyond.specguard.common.util.TokenResponseWriter;
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,7 +20,12 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
@@ -30,7 +36,6 @@ public class AuthController {
 
     private final SignupService signupService;
     private final ReissueService reissueService;
-    private final JwtUtil jwtUtil;
     private final LogoutService logoutService;
     private final InviteSignupService inviteSignupService;
     private final TokenResponseWriter tokenResponseWriter;
@@ -55,7 +60,7 @@ public class AuthController {
             HttpServletResponse response
     ) {
         String refreshToken = CookieUtil.getCookieValue(request, "refresh_token");
-        ReissueResponseDto dto= reissueService.reissue(refreshToken);
+        ReissueResponseDto dto= reissueService.reissue(false, refreshToken);
         tokenResponseWriter.writeTokens(response, dto);
         return ResponseEntity.ok().build();
