+++ backend/src/main/java/com/beyond/specguard/auth/model/filter/ClientLoginFilter.java
@@ -1,7 +1,7 @@
 package com.beyond.specguard.auth.model.filter;
 
 import com.beyond.specguard.auth.model.dto.request.LoginRequestDto;
-import com.beyond.specguard.common.util.JwtUtil;
+import com.beyond.specguard.auth.model.token.ClientAuthenticationToken;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,20 +11,13 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
-@Component
 public class ClientLoginFilter extends UsernamePasswordAuthenticationFilter {
 
-    private final AuthenticationManager clientAuthenticationManager;
-    private final JwtUtil jwtUtil;
-
-    public ClientLoginFilter(@Qualifier("clientAuthenticationManager") AuthenticationManager clientAuthenticationManager, JwtUtil jwtUtil) {
+    public ClientLoginFilter(@Qualifier("clientAuthenticationManager") AuthenticationManager clientAuthenticationManager) {
         super.setAuthenticationManager(clientAuthenticationManager);
-        this.clientAuthenticationManager = clientAuthenticationManager;
-        this.jwtUtil = jwtUtil;
         setFilterProcessesUrl("/api/v1/auth/login"); // 로그인 엔드포인트
     }
